NPM-1: Organisation and helper function to write results to markdown.
